stage 2.3.2 Added method for save Index and Lemma

+++ src/main/java/searchengine/model/Index.java
@@ -5,8 +5,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.util.HashSet;
-import java.util.Set;
 
 @Entity
 @NoArgsConstructor
@@ -29,5 +27,5 @@ public class Index {
     private Lemma lemma;
 
     @Column(name = "rank")
-    private Double rank;
+    private Integer rank;
 }
